refactor(project): split main app.rs functions into individual modules
