+++ src/ui/app.rs
@@ -113,8 +113,15 @@ use windows_sys::Win32::UI::WindowsAndMessaging::{
     SystemParametersInfoW, SPI_GETCLIENTAREAANIMATION,
 };
 
+mod list_removal;
+mod navigation_state;
 mod pages;
+mod process_refresh;
+mod runtime;
+mod settings_io;
 mod shared;
+mod tray_state;
+mod update_check;
 
 use pages::*;
 use shared::*;
@@ -956,1469 +963,7 @@ impl WinderustApp {
         app.schedule_tick(window, cx);
         app
     }
-
-    fn check_for_updates(&mut self, manual: bool, cx: &mut Context<Self>) {
-        if self.update_check_in_progress {
-            return;
-        }
-        self.update_check_in_progress = true;
-        self.update_check_message = None;
-        if manual {
-            cx.notify();
-        }
-        let channel = self.settings.general.update_channel;
-        let check = cx
-            .background_executor()
-            .spawn(async move { update_checker::check(channel) });
-        cx.spawn(async move |this, cx| {
-            let result = check.await;
-            let _ = this.update(cx, |app, cx| {
-                app.update_check_in_progress = false;
-                match result {
-                    Ok(check) => {
-                        app.latest_version = Some(check.latest_version);
-                        app.available_update = check.available_update;
-                    }
-                    Err(()) if manual => {
-                        app.update_check_message =
-                            Some(t!("about.update_check_failed").to_string());
-                    }
-                    Err(()) => {}
-                }
-                cx.notify();
-            });
-        })
-        .detach();
-    }
-
-    fn navigate_to(&mut self, page: Page, cx: &mut Context<Self>) {
-        if self.page == page {
-            return;
-        }
-
-        clear_page_hovered();
-        Self::push_navigation_page(&mut self.back_stack, self.page);
-        self.begin_breadcrumb_transition(self.page, page);
-        self.page = page;
-        self.forward_stack.clear();
-        cx.notify();
-    }
-
-    fn navigate_back(&mut self, cx: &mut Context<Self>) {
-        let Some(page) = self.back_stack.pop() else {
-            return;
-        };
-
-        clear_page_hovered();
-        Self::push_navigation_page(&mut self.forward_stack, self.page);
-        self.begin_breadcrumb_transition(self.page, page);
-        self.page = page;
-        cx.notify();
-    }
-
-    fn navigate_forward(&mut self, cx: &mut Context<Self>) {
-        let Some(page) = self.forward_stack.pop() else {
-            return;
-        };
-
-        clear_page_hovered();
-        Self::push_navigation_page(&mut self.back_stack, self.page);
-        self.begin_breadcrumb_transition(self.page, page);
-        self.page = page;
-        cx.notify();
-    }
-
-    fn begin_breadcrumb_transition(&mut self, previous: Page, current: Page) {
-        if previous == current || !ui_animations_enabled() {
-            self.breadcrumb_transition = None;
-            return;
-        }
-
-        let previous = breadcrumb_trail(previous);
-        let current = breadcrumb_trail(current);
-        if previous == current {
-            self.breadcrumb_transition = None;
-            return;
-        }
-
-        self.page_transition_generation = self.page_transition_generation.wrapping_add(1);
-        self.breadcrumb_transition = Some(BreadcrumbTransition {
-            previous,
-            current,
-            started: Instant::now(),
-            generation: self.page_transition_generation,
-        });
-    }
-
-    fn clear_finished_breadcrumb_transition(&mut self) {
-        if !ui_animations_enabled()
-            || self
-                .breadcrumb_transition
-                .as_ref()
-                .is_some_and(|transition| {
-                    transition.started.elapsed() >= Duration::from_secs_f64(MOTION_FAST_SECONDS)
-                })
-        {
-            self.breadcrumb_transition = None;
-        }
-    }
-
-    fn active_breadcrumb_transition(&self, page: Page) -> Option<&BreadcrumbTransition> {
-        self.breadcrumb_transition
-            .as_ref()
-            .filter(|transition| transition.current == breadcrumb_trail(page))
-    }
-
-    fn page_header(&self, page: Page, cx: &mut Context<Self>) -> gpui::Div {
-        page_header_with_help(
-            page,
-            self.page_header_help(page),
-            self.active_breadcrumb_transition(page),
-            cx,
-        )
-    }
-
-    fn page_header_help(&self, page: Page) -> Option<SharedString> {
-        match page {
-            Page::ActionLog => Some(action_log_page_help()),
-            _ => None,
-        }
-    }
-
-    fn page_shell(&self, _page: Page, _cx: &mut Context<Self>) -> gpui::Div {
-        page_body_shell()
-    }
-
-    fn animated_list_item(
-        &self,
-        target: ListItemRemovalTarget,
-        id: impl Into<SharedString>,
-        child: AnyElement,
-    ) -> AnyElement {
-        animated_list_item_child(
-            id,
-            child,
-            self.pending_list_item_removals.contains_key(&target),
-        )
-    }
-
-    fn request_list_item_removal(&mut self, target: ListItemRemovalTarget, cx: &mut Context<Self>) {
-        if !ui_animations_enabled() {
-            self.commit_list_item_removal(target);
-            cx.notify();
-            return;
-        }
-
-        if self.pending_list_item_removals.contains_key(&target) {
-            cx.notify();
-            return;
-        }
-
-        self.pending_list_item_removals
-            .insert(target, Instant::now());
-
-        cx.spawn(async move |this, cx| {
-            Timer::after(Duration::from_secs_f64(MOTION_FAST_SECONDS)).await;
-            let _ = this.update(cx, |app, cx| {
-                app.finish_due_list_item_removals();
-                cx.notify();
-            });
-        })
-        .detach();
-        cx.notify();
-    }
-
-    fn finish_due_list_item_removals(&mut self) {
-        let now = Instant::now();
-        let mut due = self
-            .pending_list_item_removals
-            .iter()
-            .filter_map(|(target, started)| {
-                (now.duration_since(*started) >= Duration::from_secs_f64(MOTION_FAST_SECONDS))
-                    .then_some(*target)
-            })
-            .collect::<Vec<_>>();
-
-        due.sort_by(|a, b| a.kind.cmp(&b.kind).then_with(|| b.index().cmp(&a.index())));
-
-        for target in due {
-            if self.pending_list_item_removals.remove(&target).is_some() {
-                self.commit_list_item_removal(target);
-                self.shift_pending_list_item_removals_after(target);
-            }
-        }
-    }
-
-    fn shift_pending_list_item_removals_after(&mut self, removed: ListItemRemovalTarget) {
-        let mut shifted = HashMap::new();
-        for (target, started) in self.pending_list_item_removals.drain() {
-            let target = if target.same_list(removed) && target.index() > removed.index() {
-                target.with_index(target.index() - 1)
-            } else {
-                target
-            };
-            shifted.insert(target, started);
-        }
-        self.pending_list_item_removals = shifted;
-    }
-
-    fn commit_list_item_removal(&mut self, target: ListItemRemovalTarget) {
-        let index = target.index();
-
-        match target.kind {
-            ListItemRemovalKind::ByForegroundRule => {
-                if index < self.settings.by_foreground.rules.len() {
-                    self.settings.by_foreground.rules.remove(index);
-                }
-                self.editing_rule_title = None;
-                self.expanded_rule_cards.clear();
-            }
-            ListItemRemovalKind::ByTimeRule => {
-                if index < self.settings.by_time.rules.len() {
-                    self.settings.by_time.rules.remove(index);
-                }
-                self.editing_rule_title = None;
-                self.expanded_rule_cards.clear();
-            }
-            ListItemRemovalKind::ByCpuLoadRule => {
-                if index < self.settings.by_cpu_load.rules.len() {
-                    self.settings.by_cpu_load.rules.remove(index);
-                }
-                self.editing_rule_title = None;
-                self.expanded_rule_cards.clear();
-            }
-            ListItemRemovalKind::BackgroundEfficiencyExclusion => {
-                if index < self.settings.background_efficiency.custom_rules.len() {
-                    self.settings
-                        .background_efficiency
-                        .custom_rules
-                        .remove(index);
-                }
-            }
-            ListItemRemovalKind::AppSuspensionRule => {
-                if let Some(rule) = self.settings.app_suspension.suspendable_apps.get(index) {
-                    self.expanded_rule_cards
-                        .remove(&RuleCardTarget::AppSuspension(rule.process_name.clone()));
-                }
-                if index < self.settings.app_suspension.suspendable_apps.len() {
-                    self.settings.app_suspension.suspendable_apps.remove(index);
-                }
-            }
-            ListItemRemovalKind::BackgroundCpuExclusion => {
-                if index < self.settings.background_cpu_restriction.exclusions.len() {
-                    self.settings
-                        .background_cpu_restriction
-                        .exclusions
-                        .remove(index);
-                }
-            }
-            ListItemRemovalKind::CoreLimiterRule => {
-                if let Some(rule) = self.settings.core_limiter.rules.get(index) {
-                    self.expanded_rule_cards
-                        .remove(&RuleCardTarget::CoreLimiter(rule.process_name.clone()));
-                }
-                if index < self.settings.core_limiter.rules.len() {
-                    self.settings.core_limiter.rules.remove(index);
-                }
-            }
-            ListItemRemovalKind::ByRunningAppRule => {
-                if index < self.settings.by_running_app.rules.len() {
-                    self.settings.by_running_app.rules.remove(index);
-                }
-                self.editing_rule_title = None;
-                self.expanded_rule_cards.clear();
-            }
-            ListItemRemovalKind::WorkloadEngineExclusion => {
-                if index
-                    < self
-                        .settings
-                        .workload_engine
-                        .workload_engine_exclusions
-                        .len()
-                {
-                    self.settings
-                        .workload_engine
-                        .workload_engine_exclusions
-                        .remove(index);
-                }
-            }
-            ListItemRemovalKind::ProcessPriorityExclusion => {
-                if index < self.settings.process_priority.exclusions.len() {
-                    self.settings.process_priority.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::ThreadPriorityExclusion => {
-                if index < self.settings.thread_priority.exclusions.len() {
-                    self.settings.thread_priority.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::DynamicPriorityBoostExclusion => {
-                if index < self.settings.dynamic_priority_boost.exclusions.len() {
-                    self.settings
-                        .dynamic_priority_boost
-                        .exclusions
-                        .remove(index);
-                }
-            }
-            ListItemRemovalKind::IoPriorityExclusion => {
-                if index < self.settings.io_priority.exclusions.len() {
-                    self.settings.io_priority.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::GpuPriorityExclusion => {
-                if index < self.settings.gpu_priority.exclusions.len() {
-                    self.settings.gpu_priority.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::MemoryPriorityExclusion => {
-                if index < self.settings.memory_priority.exclusions.len() {
-                    self.settings.memory_priority.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::TimerResolutionRule => {
-                if index < self.settings.timer_resolution.rules.len() {
-                    self.settings.timer_resolution.rules.remove(index);
-                }
-            }
-            ListItemRemovalKind::MemoryTrimExclusion => {
-                if index < self.settings.memory_trim.exclusions.len() {
-                    self.settings.memory_trim.exclusions.remove(index);
-                }
-            }
-            ListItemRemovalKind::CoreSteeringRule => {
-                if let Some(rule) = self.settings.core_steering.rules.get(index) {
-                    self.expanded_rule_cards
-                        .remove(&RuleCardTarget::CoreSteering(rule.process_name.clone()));
-                }
-                if index < self.settings.core_steering.rules.len() {
-                    self.settings.core_steering.rules.remove(index);
-                }
-            }
-        }
-    }
-
-    fn push_navigation_page(stack: &mut Vec<Page>, page: Page) {
-        if stack.last().copied() == Some(page) {
-            return;
-        }
-
-        stack.push(page);
-        if stack.len() > NAV_HISTORY_LIMIT {
-            stack.remove(0);
-        }
-    }
-
-    fn schedule_tick(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        let tick_interval = app_tick_interval(&self.saved_settings, self.start_minimized_applied);
-        self._tick_task = cx.spawn_in(window, async move |this, cx| {
-            Timer::after(tick_interval).await;
-            let _ = cx.update(move |window, app_cx| {
-                if let Some(this) = this.upgrade() {
-                    this.update(app_cx, |app, cx| match app.tick(window, cx) {
-                        TickOutcome::Continue { changed } => {
-                            app.schedule_tick(window, cx);
-                            if changed {
-                                cx.notify();
-                            }
-                        }
-                        TickOutcome::Stop => {}
-                    });
-                }
-            });
-        });
-    }
-
-    fn refresh_power_plans(&mut self) {
-        match self.power.list_plans() {
-            Ok(plans) => {
-                self.plans = plans;
-                self.current_plan = self.plans.iter().find(|plan| plan.active).cloned();
-                self.next_active_plan_refresh = Instant::now() + ACTIVE_PLAN_REFRESH_INTERVAL;
-                self.status_message =
-                    t!("status.loaded_power_plans", count = self.plans.len()).to_string();
-                self.ensure_processor_power_target_plan();
-                self.sync_processor_power_values_from_target_plan(false);
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn refresh_active_plan(&mut self) {
-        self.next_active_plan_refresh = Instant::now() + ACTIVE_PLAN_REFRESH_INTERVAL;
-
-        match self.power.active_plan() {
-            Ok(active) => {
-                if let Some(active) = active {
-                    let active_guid = active.guid.clone();
-                    for plan in &mut self.plans {
-                        plan.active = plan.guid.eq_ignore_ascii_case(&active_guid);
-                    }
-                    self.current_plan = self
-                        .plans
-                        .iter()
-                        .find(|plan| plan.guid.eq_ignore_ascii_case(&active_guid))
-                        .cloned()
-                        .or(Some(active));
-                    self.ensure_processor_power_target_plan();
-                    self.sync_processor_power_values_from_target_plan(false);
-                }
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn refresh_effective_power_mode(&mut self) -> bool {
-        let Some(monitor) = &self.effective_power_mode_monitor else {
-            return false;
-        };
-        let mode = monitor.snapshot();
-        if self.effective_power_mode == mode {
-            return false;
-        }
-
-        self.effective_power_mode = mode;
-        true
-    }
-
-    fn sync_adaptive_engine(&self, settings: &Settings) {
-        if settings.adaptive_engine.enabled {
-            let _ = self_power::enable_adaptive_engine();
-        } else {
-            let _ = self_power::disable_adaptive_engine();
-        }
-    }
-
-    fn run_check(&mut self, sample_dashboard: bool, now: Instant) {
-        if now >= self.next_active_plan_refresh {
-            self.refresh_active_plan();
-        }
-
-        let decision_settings = self.cached_runtime_settings();
-        let decision_settings = decision_settings.as_ref();
-        self.activity = self.activity_snapshot(decision_settings, now);
-        if sample_dashboard && self.page == Page::Home {
-            self.refresh_dashboard_resource_samples();
-        } else if decision_settings.by_cpu_load.enabled {
-            self.refresh_cpu_usage_sample(now);
-        }
-        self.foreground_app = foreground_lookup_required(decision_settings)
-            .then(|| self.foreground_detector.process_name())
-            .flatten();
-        let by_time = self
-            .by_time_scheduler
-            .current_decision(&decision_settings.by_time);
-        let by_cpu_load = self
-            .by_cpu_load_scheduler
-            .current_decision(&decision_settings.by_cpu_load, self.cpu_usage.percent);
-        self.next_schedule = self
-            .by_time_scheduler
-            .next_switch_label(&decision_settings.by_time);
-
-        self.decision = self.decision_engine.decide(
-            decision_settings,
-            DecisionInput {
-                activity_state: self.activity.state,
-                foreground_app: self.foreground_app.clone(),
-                plugged_in: power_source::is_plugged_in(),
-                by_running_app: by_running_app_decision(&self.by_running_app_status),
-                by_time,
-                by_cpu_load,
-            },
-        );
-
-        if !(decision_settings.general.enabled
-            && decision_settings.adaptive_engine.enabled
-            && decision_settings.adaptive_engine.processor_policy_enabled)
-        {
-            self.apply_decision();
-        }
-    }
-
-    fn run_check_changed(&mut self, now: Instant) -> bool {
-        let activity_state = self.activity.state;
-        let activity_idle_for = self.activity.idle_for;
-        let cpu_usage = self.cpu_usage;
-        let memory_usage = self.memory_usage;
-        let io_usage = self.io_usage;
-        let network_usage = self.network_usage;
-        let decision_target_guid = self.decision.target_guid.take();
-        let decision_state = self.decision.state;
-        let decision_reason = std::mem::take(&mut self.decision.reason);
-        let next_schedule = std::mem::take(&mut self.next_schedule);
-        let plan_count = self.plans.len();
-        let previous_active_plan_guid = active_plan_guid(&self.plans).map(str::to_owned);
-        let current_plan_guid = self.current_plan.as_ref().map(|plan| plan.guid.clone());
-        let processor_power_target_plan_personality = self.processor_power_target_plan_personality;
-        let status_message = self.status_message.clone();
-
-        self.run_check(false, now);
-
-        let resource_samples_changed = self.cpu_usage != cpu_usage
-            || self.memory_usage != memory_usage
-            || self.io_usage != io_usage
-            || self.network_usage != network_usage;
-        let resource_samples_visible = self.page == Page::Home;
-
-        self.activity.state != activity_state
-            || self.activity.idle_for != activity_idle_for
-            || (resource_samples_visible && resource_samples_changed)
-            || self.decision.target_guid != decision_target_guid
-            || self.decision.state != decision_state
-            || self.decision.reason != decision_reason
-            || self.next_schedule != next_schedule
-            || self.plans.len() != plan_count
-            || active_plan_guid(&self.plans) != previous_active_plan_guid.as_deref()
-            || self.current_plan.as_ref().map(|plan| plan.guid.as_str())
-                != current_plan_guid.as_deref()
-            || self.processor_power_target_plan_personality
-                != processor_power_target_plan_personality
-            || self.status_message != status_message
-    }
-
-    fn activity_snapshot(&mut self, settings: &Settings, now: Instant) -> ActivitySnapshot {
-        let idle_timeout = Duration::from_secs(settings.by_activity.idle_timeout_seconds);
-        let snapshot = self.idle_detector.snapshot(idle_timeout);
-        let controller_idle_for = if settings.by_activity.input_detection.controller {
-            self.controller_activity_detector.poll(now);
-            self.controller_activity_detector.idle_for(now)
-        } else {
-            self.controller_activity_detector.clear();
-            None
-        };
-
-        merge_activity_snapshot(snapshot, controller_idle_for, idle_timeout)
-    }
-
-    fn refresh_cpu_usage_sample(&mut self, now: Instant) -> bool {
-        if !refresh_due(
-            now,
-            &mut self.next_cpu_usage_refresh,
-            CPU_USAGE_REFRESH_INTERVAL,
-        ) {
-            return false;
-        }
-
-        let previous_cpu_usage = self.cpu_usage;
-        self.cpu_usage = self.cpu_monitor.sample_usage();
-        self.cpu_usage != previous_cpu_usage
-    }
-
-    fn refresh_dashboard_resource_samples(&mut self) -> bool {
-        let now = Instant::now();
-        if !refresh_due(
-            now,
-            &mut self.next_cpu_usage_refresh,
-            CPU_USAGE_REFRESH_INTERVAL,
-        ) {
-            return false;
-        }
-
-        let previous_cpu_usage = self.cpu_usage;
-        let previous_memory_usage = self.memory_usage;
-        let sample_io = refresh_due(
-            now,
-            &mut self.next_dashboard_io_refresh,
-            DASHBOARD_IO_REFRESH_INTERVAL,
-        );
-
-        self.cpu_usage = self.cpu_monitor.sample();
-        self.memory_usage = self.memory_monitor.sample();
-
-        let mut changed =
-            self.cpu_usage != previous_cpu_usage || self.memory_usage != previous_memory_usage;
-
-        if let Some(percent) = self.cpu_usage.percent {
-            if self.cpu_usage_history.len() == CPU_USAGE_HISTORY_LEN {
-                self.cpu_usage_history.pop_front();
-            }
-            self.cpu_usage_history.push_back(CpuUsageHistorySample {
-                percent: percent.clamp(0.0, 100.0),
-                frequency_mhz: self.cpu_usage.frequency_mhz,
-            });
-            changed = true;
-        }
-        if let Some(percent) = self.memory_usage.percent {
-            if self.memory_usage_history.len() == CPU_USAGE_HISTORY_LEN {
-                self.memory_usage_history.pop_front();
-            }
-            self.memory_usage_history
-                .push_back(MemoryUsageHistorySample {
-                    usage_percent: percent.clamp(0.0, 100.0),
-                    cache_percent: memory_cache_percent(self.memory_usage).unwrap_or(0.0),
-                });
-            changed = true;
-        }
-        if sample_io {
-            let previous_io_usage = self.io_usage;
-            let previous_network_usage = self.network_usage;
-            self.io_usage = self.io_monitor.sample();
-            self.network_usage = self.network_monitor.sample();
-            changed |=
-                self.io_usage != previous_io_usage || self.network_usage != previous_network_usage;
-
-            if self.io_usage.bytes_per_second.is_some() {
-                if self.io_usage_history.len() == CPU_USAGE_HISTORY_LEN {
-                    self.io_usage_history.pop_front();
-                }
-                self.io_usage_history.push_back(IoUsageHistorySample {
-                    read_bytes_per_second: self
-                        .io_usage
-                        .read_bytes_per_second
-                        .unwrap_or(0.0)
-                        .clamp(0.0, f32::MAX as f64)
-                        as f32,
-                    write_bytes_per_second: self
-                        .io_usage
-                        .write_bytes_per_second
-                        .unwrap_or(0.0)
-                        .clamp(0.0, f32::MAX as f64)
-                        as f32,
-                });
-                changed = true;
-            }
-            if self.network_usage.bytes_per_second.is_some() {
-                if self.network_usage_history.len() == CPU_USAGE_HISTORY_LEN {
-                    self.network_usage_history.pop_front();
-                }
-                self.network_usage_history
-                    .push_back(NetworkUsageHistorySample {
-                        download_bytes_per_second: self
-                            .network_usage
-                            .download_bytes_per_second
-                            .unwrap_or(0.0)
-                            .clamp(0.0, f32::MAX as f64)
-                            as f32,
-                        upload_bytes_per_second: self
-                            .network_usage
-                            .upload_bytes_per_second
-                            .unwrap_or(0.0)
-                            .clamp(0.0, f32::MAX as f64)
-                            as f32,
-                    });
-                changed = true;
-            }
-        }
-
-        changed
-    }
-
-    fn install_input_hook(&mut self, config: InputHookConfig) {
-        match InputHook::install(config, self.background_automation.input_event_callback()) {
-            Ok(input_hook) => {
-                self.input_hook = Some(input_hook);
-            }
-            Err(err) => {
-                self.status_message = err;
-            }
-        }
-    }
-
-    fn sync_input_hook(&mut self) {
-        if input_hook_required(&self.saved_settings) {
-            let config = input_hook_config(&self.saved_settings);
-            if self
-                .input_hook
-                .as_ref()
-                .is_none_or(|input_hook| input_hook.config() != config)
-            {
-                self.input_hook = None;
-                self.install_input_hook(config);
-            }
-        } else {
-            self.input_hook = None;
-        }
-    }
-
-    fn apply_decision(&mut self) {
-        let Some(target_guid) = self.decision.target_guid.as_deref() else {
-            return;
-        };
-
-        let already_active = self
-            .current_plan
-            .as_ref()
-            .is_some_and(|plan| plan.guid.eq_ignore_ascii_case(target_guid));
-        if already_active {
-            return;
-        }
-
-        if let Some((last_guid, attempted_at)) = &self.last_switch_attempt {
-            if last_guid.eq_ignore_ascii_case(target_guid)
-                && attempted_at.elapsed() < SWITCH_RETRY_INTERVAL
-            {
-                return;
-            }
-        }
-
-        self.last_switch_attempt = Some((target_guid.to_owned(), Instant::now()));
-
-        match self.power.set_active(target_guid) {
-            Ok(()) => {
-                self.status_message =
-                    t!("status.switched_power_plan", reason = self.decision.reason).to_string();
-                self.refresh_power_plans();
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn save_settings(&mut self) -> bool {
-        match config::storage::save(&self.settings) {
-            Ok(()) => {
-                self.saved_settings = self.settings.clone();
-                self.sync_input_hook();
-                self.sync_background_settings();
-                self.status_message = match startup::set_startup_with_windows(
-                    self.saved_settings.general.startup_with_windows,
-                ) {
-                    Ok(()) => t!(
-                        "status.saved_settings",
-                        path = config::storage::config_path().display()
-                    )
-                    .to_string(),
-                    Err(err) => t!("status.saved_settings_with_error", error = err).to_string(),
-                };
-                true
-            }
-            Err(err) => {
-                self.status_message = err;
-                false
-            }
-        }
-    }
-
-    fn export_settings_toml(&mut self) {
-        match choose_settings_file(self.hwnd, FileDialogMode::Save) {
-            Ok(Some(path)) => match config::storage::export_toml_to(&path, &self.settings) {
-                Ok(()) => {
-                    self.status_message =
-                        t!("status.exported_settings", path = path.display()).to_string();
-                }
-                Err(err) => self.status_message = err,
-            },
-            Ok(None) => {
-                self.status_message = t!("status.export_canceled").to_string();
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn export_action_log_csv(&mut self) {
-        if self.action_log_entries.is_empty() {
-            self.status_message = t!("status.action_log_export_empty").to_string();
-            return;
-        }
-
-        match choose_action_log_export_file(self.hwnd) {
-            Ok(Some(path)) => {
-                match fs::write(
-                    &path,
-                    action_log_entries_to_csv(self.action_log_entries.as_slice()),
-                ) {
-                    Ok(()) => {
-                        self.status_message =
-                            t!("status.exported_action_log", path = path.display()).to_string();
-                    }
-                    Err(err) => {
-                        self.status_message = t!(
-                            "status.action_log_export_failed",
-                            path = path.display(),
-                            error = err
-                        )
-                        .to_string();
-                    }
-                }
-            }
-            Ok(None) => {
-                self.status_message = t!("status.action_log_export_canceled").to_string();
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn import_settings_toml(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        match choose_settings_file(self.hwnd, FileDialogMode::Open) {
-            Ok(Some(path)) => match config::storage::import_toml_from(&path) {
-                Ok(settings) => {
-                    self.settings = settings;
-                    apply_language(self.settings.general.language);
-                    apply_appearance_settings(&self.settings.general, window, cx);
-                    match config::storage::save(&self.settings) {
-                        Ok(()) => {
-                            self.saved_settings = self.settings.clone();
-                            self.status_message = match startup::set_startup_with_windows(
-                                self.saved_settings.general.startup_with_windows,
-                            ) {
-                                Ok(()) => t!("status.imported_settings", path = path.display())
-                                    .to_string(),
-                                Err(err) => t!("status.imported_settings_with_error", error = err)
-                                    .to_string(),
-                            };
-                            self.rebuild_inputs(window, cx);
-                            self.sync_input_hook();
-                            self.sync_background_settings();
-                        }
-                        Err(err) => self.status_message = err,
-                    }
-                }
-                Err(err) => self.status_message = err,
-            },
-            Ok(None) => {
-                self.status_message = t!("status.import_canceled").to_string();
-            }
-            Err(err) => self.status_message = err,
-        }
-    }
-
-    fn process_candidates_from_info(
-        &mut self,
-        processes: Vec<ProcessCandidateInfo>,
-    ) -> Vec<ProcessCandidate> {
-        processes
-            .into_iter()
-            .map(|process| {
-                let icon = process
-                    .image_path
-                    .as_deref()
-                    .and_then(|path| self.cached_process_icon(path));
-                ProcessCandidate {
-                    name: process.name,
-                    image_path: process.image_path,
-                    icon,
-                }
-            })
-            .collect()
-    }
-
-    fn cached_process_icon(&mut self, path: &Path) -> Option<Arc<Image>> {
-        if !self.process_icon_cache.contains_key(path) {
-            let icon = load_process_icon(path);
-            self.process_icon_cache.insert(path.to_path_buf(), icon);
-        }
-
-        self.process_icon_cache.get(path).and_then(Clone::clone)
-    }
-
-    fn retain_current_process_icons(
-        cache: &mut HashMap<PathBuf, Option<Arc<Image>>>,
-        candidates: &[ProcessCandidate],
-    ) {
-        if cache.is_empty() {
-            return;
-        }
-
-        let current_paths = candidates
-            .iter()
-            .filter_map(|candidate| candidate.image_path.as_deref())
-            .collect::<HashSet<_>>();
-        let old_len = cache.len();
-        cache.retain(|path, _| current_paths.contains(path.as_path()));
-        if cache.len() != old_len {
-            cache.shrink_to_fit();
-        }
-    }
-
-    fn refresh_process_candidates(&mut self, report_status: bool) -> bool {
-        self.next_process_refresh = Instant::now() + PROCESS_REFRESH_INTERVAL;
-        match list_process_candidates() {
-            Ok(processes) => {
-                let processes = self.process_candidates_from_info(processes);
-                let changed = self.process_candidates != processes;
-                self.process_candidates = processes;
-                Self::retain_current_process_icons(
-                    &mut self.process_icon_cache,
-                    &self.process_candidates,
-                );
-                if report_status {
-                    let message = t!(
-                        "status.loaded_running_apps",
-                        count = self.process_candidates.len()
-                    )
-                    .to_string();
-                    let status_changed = self.status_message != message;
-                    self.status_message = message;
-                    changed || status_changed
-                } else {
-                    changed
-                }
-            }
-            Err(err) => {
-                let changed = self.status_message != err;
-                self.status_message = err;
-                changed
-            }
-        }
-    }
-
-    fn refresh_running_processes(&mut self, report_status: bool) -> bool {
-        self.next_process_refresh = Instant::now() + PROCESS_REFRESH_INTERVAL;
-        match list_processes() {
-            Ok(mut processes) => {
-                processes.sort_by(|left, right| {
-                    left.name
-                        .cmp(&right.name)
-                        .then_with(|| left.id.cmp(&right.id))
-                });
-                let changed = self.running_processes != processes;
-                self.running_processes = processes;
-                let expanded_group_count = self.expanded_process_list_groups.len();
-                if expanded_group_count != 0 {
-                    let active_group_keys = self
-                        .running_processes
-                        .iter()
-                        .map(|process| process_list_group_key(&process.name))
-                        .collect::<HashSet<_>>();
-                    self.expanded_process_list_groups
-                        .retain(|key| active_group_keys.contains(key));
-                }
-                let groups_changed =
-                    self.expanded_process_list_groups.len() != expanded_group_count;
-                if report_status {
-                    let message = t!(
-                        "status.loaded_running_processes",
-                        count = self.running_processes.len()
-                    )
-                    .to_string();
-                    let status_changed = self.status_message != message;
-                    self.status_message = message;
-                    changed || groups_changed || status_changed
-                } else {
-                    changed || groups_changed
-                }
-            }
-            Err(err) => {
-                let changed = self.status_message != err;
-                self.status_message = err;
-                changed
-            }
-        }
-    }
-
-    fn sync_tray_icon(&mut self) -> bool {
-        let tray_required =
-            self.settings.general.hide_to_tray || self.saved_settings.general.start_minimized;
-        let tray_present = self.tray_icon.is_some();
-        let mut changed = false;
-
-        if tray_required {
-            if self.tray_icon.is_none() {
-                let Some(hwnd) = self.hwnd else {
-                    self.set_tray_hide_on_close(false);
-                    let message = t!("status.system_tray_unavailable").to_string();
-                    if self.status_message != message {
-                        self.status_message = message;
-                        changed = true;
-                    }
-                    return changed;
-                };
-
-                match TrayIcon::install(hwnd) {
-                    Ok(icon) => {
-                        self.tray_icon = Some(icon);
-                        changed = true;
-                        let message = t!("status.system_tray_enabled").to_string();
-                        if self.status_message != message {
-                            self.status_message = message;
-                            changed = true;
-                        }
-                    }
-                    Err(err) => {
-                        if self.status_message != err {
-                            self.status_message = err;
-                            changed = true;
-                        }
-                    }
-                }
-            }
-            self.set_tray_hide_on_close(
-                self.settings.general.hide_to_tray && self.tray_icon.is_some(),
-            );
-        } else if self.tray_icon.take().is_some() {
-            self.set_tray_hide_on_close(false);
-            changed = true;
-            let message = t!("status.system_tray_disabled").to_string();
-            if self.status_message != message {
-                self.status_message = message;
-                changed = true;
-            }
-        } else {
-            self.set_tray_hide_on_close(false);
-        }
-
-        changed || tray_present != self.tray_icon.is_some()
-    }
-
-    fn set_tray_hide_on_close(&mut self, enabled: bool) {
-        if self.tray_hide_on_close == enabled {
-            return;
-        }
-
-        self.tray_hide_on_close = enabled;
-        tray::set_hide_on_close(enabled);
-    }
-
-    fn apply_start_minimized(&mut self, window: &mut Window) -> bool {
-        if self.start_minimized_applied {
-            return false;
-        }
-        self.start_minimized_applied = true;
-
-        if !self.saved_settings.general.start_minimized {
-            return false;
-        }
-
-        if self.tray_icon.is_some() {
-            if let Some(hwnd) = self.hwnd {
-                tray::hide_window(hwnd);
-                self.status_message = t!("status.started_in_tray").to_string();
-                return true;
-            }
-        }
-
-        window.minimize_window();
-        self.status_message = t!("status.started_minimized").to_string();
-        true
-    }
-
-    fn refresh_after_tray_restore(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        self.next_check = Instant::now();
-        match self.tick(window, cx) {
-            TickOutcome::Continue { changed } => {
-                self.schedule_tick(window, cx);
-                if changed {
-                    cx.notify();
-                }
-            }
-            TickOutcome::Stop => {}
-        }
-    }
-
-    fn tick(&mut self, window: &mut Window, cx: &mut Context<Self>) -> TickOutcome {
-        if tray::take_quit_requested() {
-            self.set_tray_hide_on_close(false);
-            self.tray_icon = None;
-            window.remove_window();
-            return TickOutcome::Stop;
-        }
-
-        let mut changed = self.apply_start_minimized(window);
-        changed |= self.apply_pending_auto_exclusions();
-        if tray::is_hidden_to_tray() {
-            self.sync_input_hook();
-            self.sync_background_settings();
-            return TickOutcome::Stop;
-        }
-
-        if let Some(background_status) = self
-            .background_automation
-            .status_snapshot_since(self.last_background_status_generation)
-        {
-            self.last_background_status_generation = background_status.generation;
-
-            if self.background_efficiency_status != background_status.background_efficiency {
-                self.background_efficiency_status = background_status.background_efficiency;
-                changed = true;
-            }
-
-            if self.app_suspension_status != background_status.app_suspension {
-                self.app_suspension_status = background_status.app_suspension;
-                changed = true;
-            }
-
-            if self.core_limiter_status != background_status.core_limiter {
-                self.core_limiter_status = background_status.core_limiter;
-                changed = true;
-            }
-
-            if self.core_steering_status != background_status.core_steering {
-                self.core_steering_status = background_status.core_steering;
-                changed = true;
-            }
-
-            if self.background_cpu_restriction_status
-                != background_status.background_cpu_restriction
-            {
-                self.background_cpu_restriction_status =
-                    background_status.background_cpu_restriction;
-                changed = true;
-            }
-
-            if self.by_running_app_status != background_status.by_running_app {
-                self.by_running_app_status = background_status.by_running_app;
-                changed = true;
-            }
-
-            if self.workload_engine_status != background_status.workload_engine {
-                self.workload_engine_status = background_status.workload_engine;
-                changed = true;
-            }
-
-            if self.process_priority_status != background_status.process_priority {
-                self.process_priority_status = background_status.process_priority;
-                changed = true;
-            }
-
-            if self.thread_priority_status != background_status.thread_priority {
-                self.thread_priority_status = background_status.thread_priority;
-                changed = true;
-            }
-
-            if self.dynamic_priority_boost_status != background_status.dynamic_priority_boost {
-                self.dynamic_priority_boost_status = background_status.dynamic_priority_boost;
-                changed = true;
-            }
-
-            if self.io_priority_status != background_status.io_priority {
-                self.io_priority_status = background_status.io_priority;
-                changed = true;
-            }
-
-            if self.gpu_priority_status != background_status.gpu_priority {
-                self.gpu_priority_status = background_status.gpu_priority;
-                changed = true;
-            }
-
-            if self.memory_priority_status != background_status.memory_priority {
-                self.memory_priority_status = background_status.memory_priority;
-                changed = true;
-            }
-
-            if self.memory_trim_status != background_status.memory_trim {
-                self.memory_trim_status = background_status.memory_trim;
-                changed = true;
-            }
-
-            if self.timer_resolution_status != background_status.timer_resolution {
-                self.timer_resolution_status = background_status.timer_resolution;
-                changed = true;
-            }
-
-            if !Arc::ptr_eq(
-                &self.action_log_entries,
-                &background_status.action_log_entries,
-            ) {
-                self.action_log_entries = background_status.action_log_entries;
-                changed = true;
-            }
-
-            if self.last_appearance_change_generation
-                != background_status.appearance_change_generation
-            {
-                self.last_appearance_change_generation =
-                    background_status.appearance_change_generation;
-                apply_appearance_settings(&self.settings.general, window, cx);
-                changed = true;
-            }
-        }
-
-        changed |= self.refresh_effective_power_mode();
-
-        let now = Instant::now();
-
-        if self.page == Page::TimerResolution
-            && !self.settings.timer_resolution.enabled
-            && refresh_due(
-                now,
-                &mut self.next_timer_resolution_status_refresh,
-                TIMER_RESOLUTION_STATUS_REFRESH_INTERVAL,
-            )
-        {
-            let timer_resolution_status =
-                timer_resolution::query_snapshot(self.settings.timer_resolution.enabled);
-            if self.timer_resolution_status != timer_resolution_status {
-                self.timer_resolution_status = timer_resolution_status;
-                changed = true;
-            }
-        }
-
-        if now >= self.next_process_refresh {
-            if self.page == Page::ProcessList {
-                changed |= self.refresh_running_processes(false);
-            } else if self.page_uses_process_candidates() {
-                changed |= self.refresh_process_candidates(false);
-            }
-        }
-
-        if self.page == Page::Home {
-            changed |= self.refresh_dashboard_resource_samples();
-        }
-
-        let should_check_now = now >= self.next_check;
-
-        if should_check_now {
-            changed |= self.run_check_changed(now);
-            self.next_check = now
-                + Duration::from_millis(
-                    self.settings
-                        .general
-                        .check_interval_ms
-                        .max(ACTIVITY_CHECK_INTERVAL_MIN_MS),
-                );
-        }
-
-        changed |= self.sync_tray_icon();
-
-        if !should_check_now {
-            self.sync_background_settings();
-        }
-        TickOutcome::Continue { changed }
-    }
-
-    fn apply_pending_auto_exclusions(&mut self) -> bool {
-        let Some(pending) = self
-            .background_automation
-            .take_pending_auto_exclusions_since(&mut self.last_pending_auto_exclusions_generation)
-        else {
-            return false;
-        };
-        let mut changed = false;
-
-        for process in pending.background_efficiency {
-            if can_add_background_efficiency_process(&self.settings.background_efficiency, &process)
-            {
-                self.settings
-                    .background_efficiency
-                    .custom_rules
-                    .push(new_background_efficiency_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.app_suspension {
-            if can_add_app_suspension_process(&self.settings.app_suspension, &process) {
-                let mut rule = new_app_suspension_rule(&process);
-                rule.enabled = false;
-                self.settings.app_suspension.suspendable_apps.push(rule);
-                changed = true;
-            }
-        }
-
-        for process in pending.core_steering {
-            if can_add_core_steering_process(&self.settings.core_steering, &process) {
-                let mut rule = new_core_steering_rule(&process);
-                rule.enabled = false;
-                self.settings.core_steering.rules.push(rule);
-                changed = true;
-            }
-        }
-
-        for process in pending.background_cpu_restriction {
-            if can_add_background_cpu_exclusion(&self.settings.background_cpu_restriction, &process)
-            {
-                self.settings
-                    .background_cpu_restriction
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.core_limiter {
-            if can_add_core_limiter_process(&self.settings.core_limiter, &process) {
-                let mut rule = new_core_limiter_rule(&process);
-                rule.enabled = false;
-                self.settings.core_limiter.rules.push(rule);
-                changed = true;
-            }
-        }
-
-        for process in pending.workload_engine {
-            if can_add_workload_engine_exclusion(&self.settings.workload_engine, &process) {
-                self.settings
-                    .workload_engine
-                    .workload_engine_exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.io_priority {
-            if can_add_io_priority_exclusion(&self.settings.io_priority, &process) {
-                self.settings
-                    .io_priority
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.process_priority {
-            if can_add_process_priority_exclusion(&self.settings.process_priority, &process) {
-                self.settings
-                    .process_priority
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.thread_priority {
-            if can_add_thread_priority_exclusion(&self.settings.thread_priority, &process) {
-                self.settings
-                    .thread_priority
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.dynamic_priority_boost {
-            if can_add_dynamic_priority_boost_exclusion(
-                &self.settings.dynamic_priority_boost,
-                &process,
-            ) {
-                self.settings
-                    .dynamic_priority_boost
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.gpu_priority {
-            if can_add_gpu_priority_exclusion(&self.settings.gpu_priority, &process) {
-                self.settings
-                    .gpu_priority
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.memory_priority {
-            if can_add_memory_priority_exclusion(&self.settings.memory_priority, &process) {
-                self.settings
-                    .memory_priority
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        for process in pending.memory_trim {
-            if can_add_memory_trim_exclusion(&self.settings.memory_trim, &process) {
-                self.settings
-                    .memory_trim
-                    .exclusions
-                    .push(new_process_exclusion_rule(&process));
-                changed = true;
-            }
-        }
-
-        if changed {
-            self.save_settings();
-        }
-
-        changed
-    }
-
-    fn page_uses_process_candidates(&self) -> bool {
-        matches!(
-            self.page,
-            Page::ByForeground
-                | Page::BackgroundEfficiency
-                | Page::AppSuspension
-                | Page::ProcessPriority
-                | Page::DynamicPriorityBoost
-                | Page::CoreLimiter
-                | Page::BackgroundCpuRestriction
-                | Page::IoPriority
-                | Page::GpuPriority
-                | Page::MemoryPriority
-                | Page::TimerResolution
-                | Page::ByRunningApp
-                | Page::CoreSteering
-        )
-    }
-
-    fn cancel_settings_changes(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        let had_unsaved_changes = self.settings != self.saved_settings;
-        self.settings = self.saved_settings.clone();
-        apply_language(self.settings.general.language);
-        apply_appearance_settings(&self.settings.general, window, cx);
-        self.status_message = t!("status.unsaved_canceled").to_string();
-        self.editing_rule_title = None;
-        self.expanded_rule_cards.clear();
-        self.rebuild_inputs(window, cx);
-        if had_unsaved_changes {
-            self.start_unsaved_popup_vanish();
-        }
-    }
-
-    fn start_unsaved_popup_vanish(&mut self) {
-        self.unsaved_popup_was_visible = false;
-        self.unsaved_popup_vanish_started = ui_animations_enabled().then_some(Instant::now());
-    }
-
-    fn unsaved_popup_vanish_progress(&mut self, unsaved: bool, window: &mut Window) -> Option<f32> {
-        if unsaved {
-            self.unsaved_popup_was_visible = true;
-            self.unsaved_popup_vanish_started = None;
-            return None;
-        }
-
-        if !ui_animations_enabled() {
-            self.unsaved_popup_was_visible = false;
-            self.unsaved_popup_vanish_started = None;
-            return None;
-        }
-
-        if self.unsaved_popup_vanish_started.is_none() {
-            if self.unsaved_popup_was_visible {
-                self.start_unsaved_popup_vanish();
-            } else {
-                return None;
-            }
-        } else {
-            self.unsaved_popup_was_visible = false;
-        }
-
-        let started = self.unsaved_popup_vanish_started?;
-        let duration = Duration::from_secs_f64(UNSAVED_POPUP_VANISH_SECONDS);
-        let elapsed = started.elapsed();
-        if elapsed >= duration {
-            self.unsaved_popup_was_visible = false;
-            self.unsaved_popup_vanish_started = None;
-            None
-        } else {
-            window.request_animation_frame();
-            Some(expandable_motion_ease(
-                (elapsed.as_secs_f32() / duration.as_secs_f32().max(f32::EPSILON)).clamp(0.0, 1.0),
-                false,
-            ))
-        }
-    }
-
-    fn background_settings(&self) -> Settings {
-        self.runtime_settings()
-    }
-
-    fn runtime_settings(&self) -> Settings {
-        runtime_settings_from(&self.settings, &self.saved_settings)
-    }
-
-    fn cached_runtime_settings(&mut self) -> Arc<Settings> {
-        self.sync_background_settings();
-        Arc::clone(&self.last_background_settings)
-    }
-
-    fn sync_background_settings(&mut self) {
-        if runtime_settings_matches(
-            self.last_background_settings.as_ref(),
-            &self.settings,
-            &self.saved_settings,
-        ) {
-            return;
-        }
-
-        let settings = Arc::new(self.background_settings());
-        self.sync_adaptive_engine(settings.as_ref());
-        self.background_automation
-            .update_settings(settings.as_ref());
-        self.last_background_settings = settings;
-    }
 }
-
 impl Drop for WinderustApp {
     fn drop(&mut self) {
         let _ = self_power::disable_adaptive_engine();
